requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
anthropic>=0.25.0
